@@ -20,11 +20,10 @@ requirements = [
                 'graphviz',
                 'nifty7',
                 'sympy',
-                'rich',
                 "codex-africanus[complete]",
 
                 "dask-ms[xarray, zarr, s3]"
-                "@git+https://github.com/ska-sa/dask-ms.git"
+                "@git+https://github.com/ratt-ru/dask-ms.git"
                 "@master",
 
                 "packratt"
@@ -55,7 +54,7 @@ setup(
      packages=find_packages(),
      include_package_data=True,
      zip_safe=False,
-     python_requires='>=3.7',
+     python_requires='>=3.8',
      install_requires=requirements,
      classifiers=[
          "Programming Language :: Python :: 3",
