@@ -58,14 +58,16 @@ def hogbom(
         if verbosity:
             print("Maximum iterations reached. Max of residual = %f." %
                   (IRmax), file=log)
+        return model, 1
     elif stall_count >= 5:
         if verbosity:
             print("Stalled. Max of residual = %f." %
                   (IRmax), file=log)
+        return model, 1
     else:
         if verbosity:
             print("Success, converged after %i iterations" % k, file=log)
-    return x
+        return x, 0
 
 
 # import jax.numpy as jnp
