@@ -5,6 +5,47 @@ import pyscilog
 log = pyscilog.get_logger('PCG')
 
 
+def cg(A,
+       b,
+       x0=None,
+       tol=1e-5,
+       maxit=500,
+       verbosity=1,
+       report_freq=10):
+
+    if x0 is None:
+        x = np.zeros(b.shape, dtype=b.dtype)
+    else:
+        x = x0.copy()
+
+    # initial residual
+    r = A(x) - b
+    p = -r
+    rnorm = np.vdot(r, r)
+    rnorm0 = rnorm
+    eps = rnorm
+    k = 0
+    while eps > tol and k < maxit:
+        xp = x
+        rp = r
+        Ap = A(p)
+        alpha = rnorm/np.vdot(p, Ap)
+        x += alpha * p
+        r += alpha * Ap
+        rnorm_next = np.vdot(r, r)
+        beta = rnorm_next/rnorm
+        p = beta*p - r
+        rnorm = rnorm_next
+        eps = rnorm #/rnorm0
+
+        k += 1
+
+    if k >= maxit:
+        print(f"Max iters reached eps = {eps}")
+
+    return x
+
+
 def pcg(A,
         b,
         x0=None,
@@ -14,7 +55,8 @@ def pcg(A,
         minit=100,
         verbosity=1,
         report_freq=10,
-        backtrack=True):
+        backtrack=True,
+        return_resid=False):
 
     if x0 is None:
         x0 = np.zeros(b.shape, dtype=b.dtype)
@@ -83,7 +125,10 @@ def pcg(A,
     else:
         if verbosity:
             print(f"Success, converged after {k} iterations", file=log)
-    return x
+    if not return_resid:
+        return x
+    else:
+        return x, r
 
 from pfb.operators.psf import _hessian_reg_psf as hessian_psf
 def _pcg_psf_impl(psfhat,
