@@ -66,6 +66,7 @@ def _fwdbwd(**kw):
     from pfb.opt.power_method import power_method
     from pfb.prox.prox_21m import prox_21m
     from pfb.prox.prox_21 import prox_21
+    from pfb.wavelets.wavelets import wavelet_setup
     import pywt
     from copy import deepcopy
 
@@ -124,36 +125,13 @@ def _fwdbwd(**kw):
 
     # dictionary setup
     print("Setting up dictionary", file=log)
-    bases = opts.bases.split(',')
-    ntots = []
-    iys = {}
-    sys = {}
-    for base in bases:
-        if base == 'self':
-            y, iy, sy = model[0].ravel(), 0, 0
-        else:
-            alpha = pywt.wavedecn(model[0], base, mode='zero',
-                                  level=opts.nlevels)
-            y, iy, sy = pywt.ravel_coeffs(alpha)
-        iys[base] = iy
-        sys[base] = sy
-        ntots.append(y.size)
-
-    # get padding info
-    nmax = np.asarray(ntots).max()
-    padding = []
-    nbasis = len(ntots)
-    for i in range(nbasis):
-        padding.append(slice(0, ntots[i]))
-
-
-    # initialise dictionary operators
-    bases = da.from_array(np.array(bases, dtype=object), chunks=-1)
-    ntots = da.from_array(np.array(ntots, dtype=object), chunks=-1)
-    padding = da.from_array(np.array(padding, dtype=object), chunks=-1)
-    psiH = partial(im2coef, bases=bases, ntot=ntots, nmax=nmax,
+    bases = tuple(opts.bases.split(','))
+    nbasis = len(bases)
+    iys, sys, ntot, nmax = wavelet_setup(residual, bases, opts.nlevels)
+    ntot = tuple(ntot)
+    psiH = partial(im2coef, bases=bases, ntot=ntot, nmax=nmax,
                    nlevels=opts.nlevels)
-    psi = partial(coef2im, bases=bases, padding=padding,
+    psi = partial(coef2im, bases=bases, ntot=ntot,
                   iy=iys, sy=sys, nx=nx, ny=ny)
 
     hessopts = {}
@@ -199,17 +177,29 @@ def _fwdbwd(**kw):
     else:
         weight = np.ones((nbasis, nmax), dtype=model.dtype)
 
+    # for saving intermediaries
+    ra = mds.ra
+    dec = mds.dec
+    radec = [ra, dec]
+    cell_rad = mds.cell_rad
+    cell_deg = np.rad2deg(cell_rad)
+    freq_out = mds.band.values
+    hdr = set_wcs(cell_deg, cell_deg, nx, ny, radec, freq_out)
+
     residual_mfs = np.sum(residual, axis=0)
     rms = np.std(residual_mfs)
     rmax = residual_mfs.max()
     print(f"It {0}: max resid = {rmax:.3e}, rms = {rms:.3e}", file=log)
     for i in range(opts.niter):
         print('Getting update', file=log)
-        update = pcg(hess, mask[None] * residual,
+        update, fwd_resid = pcg(hess, mask[None] * residual,
                      tol=opts.cg_tol, maxit=opts.cg_maxit,
                      minit=opts.cg_minit, verbosity=opts.cg_verbose,
                      report_freq=opts.cg_report_freq,
-                     backtrack=opts.backtrack)
+                     backtrack=opts.backtrack, return_resid=True)
+
+        save_fits(f'{basename}_update_{i}.fits', update, hdr)
+        save_fits(f'{basename}_fwd_resid_{i}.fits', fwd_resid, hdr)
 
         print('Getting model', file=log)
         modelp = deepcopy(model)
